+++ lib/market/index.ts
@@ -2,17 +2,7 @@ import Web3 from 'web3'
 import Contract from 'web3/eth/contract'
 import { marketAbi } from './abi'
 import { CustomOptions } from '../option'
-
-export const createBaseSchemaCaller = async (
-	contract: Contract
-): Promise<string> => contract.methods.schema().call()
-
-export const createSchemaCaller = (
-	contract: Contract
-): (() => Promise<string[]>) => async () =>
-	createBaseSchemaCaller(contract).then(
-		result => JSON.parse(result) as string[]
-	)
+import { createSchemaCaller } from './schema'
 
 export interface CreateMarketContract {
 	schema: () => Promise<string[]>
